rename NotificationIndicator
