+++ src/components/UI/NotificationIndicator.vue
